.h file sharing across overlays
